+++ lib/common/errors.js
@@ -1,0 +1,13 @@
+
+class BaseError extends Error {
+  constructor(message, status) {
+    super(message);
+    this.name = this.constructor.name;
+    Error.captureStackTrace(this, this.constructor);
+    this.status = status || 500;
+  }
+}
+
+module.exports = {
+  BaseError,
+};
