+++ lib/food/errors.js
@@ -1,15 +1,7 @@
 'use strict';
 
 const urls = require('../common/baseUrl');
-
-class BaseError extends Error {
-  constructor(message, status) {
-    super(message);
-    this.name = this.constructor.name;
-    Error.captureStackTrace(this, this.constructor);
-    this.status = status || 500;
-  }
-}
+const {BaseError} = require('../common/errors.js');
 
 class SsuDormConnectionError extends BaseError {
   constructor(message) {
