+++ lib/food/index.js
@@ -2,10 +2,10 @@
 
 const request = require('request');
 const haksik = require('./haksik');
-const iconv = require('iconv');
+const {Iconv} = require('iconv');
 const errors = require('./errors');
 
-const iConv = new iconv.Iconv('euc-kr', 'utf-8//translit//ignore');
+const iConv = new Iconv('euc-kr', 'utf-8//translit//ignore');
 
 async function getOfficer(dayOfWeek) {
   // params: dayOfWeek = 0~6
